compliance dash core functionality
